Fold this branch's test helpers into main's
main grew buildkite/helpers_test.go with the same three helpers this
branch had put in framework_test.go, so merging main declared each one
twice and the package no longer compiled. main's file is the superset,
so keep it and carry over updateRequestFor, the only helper it lacked.
